run end to end tests sequentially
mocha runs file simultaneously which causes some tests to fail and
necessitates the creation of new database to make these tests pass. So,
I created a custom test runner that chains promises together to ensure
that the tests are run in order. This also enables the tests to be
performed on a single database rather than creating new one.

Whats done here:
 - created a `when` that defines the test suite
 - created a `given` that sets up the test
 - created a `it` that runs the actual test

Ex:
```javascript
when('running a test')
  .then(given(() => {/*setup*/})
  .then(it('should evaluate the code', () => {/*test*/})

Some of these are hardcoded, like test file location (`test/`).

Finally there is the `executeSteps`, which takes in a list of params
that are the names of test suites in `test` folder. They all export a
function which returns a promise, built using `when`.

Other changes:
 - removed mocha from package and eslint
 - added chalk back for colorful logs :tada:
 - removed all the numbering, which was so disturbing :goberserk:
 - updated eslint globals for test runners
 - extracted `fs.readFile` to a shared promise

:v: